# docs/元数据服务/openapi.json
{
  "components": {
    "schemas": {
...
      "_classify_bind_Response": {
        "properties": {
          "retCode": {
            "format": "string",
            "description": "返回码",
            "type": "string"
          },
          "retMsg": {
            "format": "string",
            "description": "返回说明",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "提示",
            "type": "string"
          },
          "data": {
            "description": "返回数据(object)",
            "type": "object"
          }
        }
      },
...
      "_classify__ids_Response": {
        "properties": {
          "retCode": {
            "format": "string",
            "description": "返回码",
            "type": "string"
          },
          "retMsg": {
            "format": "string",
            "description": "返回说明",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "提示",
            "type": "string"
          },
          "data": {
            "description": "返回数据(object)",
            "type": "object"
          }
        }
      },
...
      "_classify_Response": {
        "properties": {
          "retCode": {
            "format": "string",
            "description": "返回码",
            "type": "string"
          },
          "retMsg": {
            "format": "string",
            "description": "返回说明",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "提示",
            "type": "string"
          },
          "data": {
            "description": "返回数据",
            "type": "array",
            "items": {
              "properties": {
                "name": {
                  "format": "string",
                  "description": "名称",
                  "type": "string"
                },
                "description": {
                  "format": "string",
                  "description": "描述",
                  "type": "string"
                },
                "id": {
                  "format": "int64",
                  "description": "No comments found.",
                  "type": "integer"
                },
                "gmtCreate": {
                  "format": "string",
                  "description": "创建时间",
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "_classify_bind_Request": {
        "properties": {
          "guid": {
            "format": "string",
            "description": "元数据标识",
            "type": "string"
          },
          "classify": {
            "format": "string",
            "description": "分类",
            "type": "string"
          }
        },
        "required": [
          "guid"
        ]
      },
...
      "_classify_Request": {},
...
      "_classify__id_Response": {
        "properties": {
          "retCode": {
            "format": "string",
            "description": "返回码",
            "type": "string"
          },
          "retMsg": {
            "format": "string",
            "description": "返回说明",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "提示",
            "type": "string"
          },
          "data": {
            "description": "返回数据(object)",
            "type": "object"
          }
        }
      },
...
      "_classify_bind__guid_Response": {
        "properties": {
          "retCode": {
            "format": "string",
            "description": "返回码",
            "type": "string"
          },
          "retMsg": {
            "format": "string",
            "description": "返回说明",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "提示",
            "type": "string"
          },
          "data": {
            "description": "返回数据(object)",
            "type": "object",
            "properties": {
              "guid": {
                "format": "string",
                "description": "元数据标识",
                "type": "string"
              },
              "classify": {
                "format": "string",
                "description": "分类",
                "type": "string"
              }
            }
          }
        }
      },
...
      "_classify__id_Request": {
        "properties": {
          "name": {
            "format": "string",
            "description": "名称",
            "type": "string"
          },
          "description": {
            "format": "string",
            "description": "描述",
            "type": "string"
          }
        }
      },
...
      "_labels_bind__guid_Response": {
        "properties": {
          "retCode": {
            "format": "string",
            "description": "返回码",
            "type": "string"
          },
          "retMsg": {
            "format": "string",
            "description": "返回说明",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "提示",
            "type": "string"
          },
          "data": {
            "description": "返回数据(object)",
            "type": "object",
            "properties": {
              "guid": {
                "format": "string",
                "description": "元数据标识",
                "type": "string"
              },
              "labels": {
                "format": "string",
                "description": "元数据标签",
                "type": "string"
              }
            }
          }
        }
      },
...
    }
  },
  "servers": [
    {
      "url": "http://172.20.0.13:8782/mtd"
    }
  ],
  "openapi": "3.0.3",
  "paths": {
    "/classify/": {
      "post": {
        "summary": "新增元数据分类",
        "deprecated": false,
        "description": "新增元数据分类",
        "tags": [
          "MtdClassifyController"
        ],
        "requestBody": {
          "content": {
            "application/json; charset\u003dutf-8": {
              "schema": {
                "$ref": "#/components/schemas/_classify_Request"
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "{\n  \"name\": \"rochel.emard\",\n  \"description\": \"m8t613\"\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/_classify_Response"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"retCode\": \"365\",\n  \"retMsg\": \"cqcoi3\",\n  \"tips\": \"flt57r\",\n  \"data\": {\n    \"waring\": \"You may have used non-display generics.\"\n  }\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "insert",
        "parameters": []
      },
      "get": {
        "summary": "查询元数据分类",
        "deprecated": false,
        "description": "查询元数据分类",
        "tags": [
          "MtdClassifyController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/_classify_Response"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"retCode\": \"365\",\n  \"retMsg\": \"9uywgv\",\n  \"tips\": \"nufxo3\",\n  \"data\": [\n    {\n      \"name\": \"rochel.emard\",\n      \"description\": \"u20pfh\",\n      \"id\": 214,\n      \"gmtCreate\": \"2021-10-25 16:17:38\"\n    }\n  ]\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "classifyByAll",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "description": "名称",
            "required": false,
            "example": "rochel.emard",
            "name": "name"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "description": "描述",
            "required": false,
            "example": "q7u6uy",
            "name": "description"
          }
        ]
      }
    },
...
    "/classify/bind/{guid}": {
      "get": {
        "summary": "查询元数据分类绑定关系",
        "deprecated": false,
        "description": "查询元数据分类绑定关系",
        "tags": [
          "MtdClassifyAtlasController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/_classify_bind__guid_Response"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"retCode\": \"365\",\n  \"retMsg\": \"vjo0cx\",\n  \"tips\": \"5gnt6e\",\n  \"data\": {\n    \"guid\": \"59\",\n    \"classify\": \"dztsfl\"\n  }\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "getByGuid",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "description": "元数据guid",
            "required": true,
            "example": "59",
            "name": "guid"
          }
        ]
      }
    },
...
    "/classify/{ids}": {
      "delete": {
        "summary": "删除元数据分类",
        "deprecated": false,
        "description": "删除元数据分类",
        "tags": [
          "MtdClassifyController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/_classify__ids_Response"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"retCode\": \"365\",\n  \"retMsg\": \"i3uln1\",\n  \"tips\": \"m8zm87\",\n  \"data\": {\n    \"waring\": \"You may have used non-display generics.\"\n  }\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "delete",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "description": "id集合，String类型",
            "required": true,
            "example": "x2ae87",
            "name": "ids"
          }
        ]
      }
    },
    "/classify/{id}": {
      "put": {
        "summary": "修改元数据分类",
        "deprecated": false,
        "description": "修改元数据分类",
        "tags": [
          "MtdClassifyController"
        ],
        "requestBody": {
          "content": {
            "application/json; charset\u003dutf-8": {
              "schema": {
                "$ref": "#/components/schemas/_classify__id_Request"
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "{\n  \"name\": \"rochel.emard\",\n  \"description\": \"q7e3nr\"\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/_classify__id_Response"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"retCode\": \"365\",\n  \"retMsg\": \"nd77gv\",\n  \"tips\": \"yiws3z\",\n  \"data\": {\n    \"waring\": \"You may have used non-display generics.\"\n  }\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "update",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "in": "path",
            "description": "No comments found.",
            "required": true,
            "example": "770",
            "name": "id"
          }
        ]
      }
    },
    "/labels/bind/{guid}": {
      "get": {
        "summary": "查询元数据标签绑定关系",
        "deprecated": false,
        "description": "查询元数据标签绑定关系",
        "tags": [
          "MtdLabelsAtlasController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/_labels_bind__guid_Response"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"retCode\": \"365\",\n  \"retMsg\": \"rs441k\",\n  \"tips\": \"fs8cw7\",\n  \"data\": {\n    \"guid\": \"59\",\n    \"labels\": \"pkfgtc\"\n  }\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "getByGuid",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "description": "元数据唯一标识",
            "required": true,
            "example": "59",
            "name": "guid"
          }
        ]
      }
    },
...
    "/classify/bind/": {
      "post": {
        "summary": "新增元数据分类绑定关系",
        "deprecated": false,
        "description": "新增元数据分类绑定关系",
        "tags": [
          "MtdClassifyAtlasController"
        ],
        "requestBody": {
          "content": {
            "application/json; charset\u003dutf-8": {
              "schema": {
                "$ref": "#/components/schemas/_classify_bind_Request"
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "{\n  \"guid\": \"59\",\n  \"classify\": \"2ypbmj\"\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/_classify_bind_Response"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"retCode\": \"365\",\n  \"retMsg\": \"5w3gtr\",\n  \"tips\": \"4ljjn8\",\n  \"data\": {\n    \"waring\": \"You may have used non-display generics.\"\n  }\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "insert",
        "parameters": []
      },
      "put": {
        "summary": "修改元数据分类绑定关系",
        "deprecated": false,
        "description": "修改元数据分类绑定关系",
        "tags": [
          "MtdClassifyAtlasController"
        ],
        "requestBody": {
          "content": {
            "application/json; charset\u003dutf-8": {
              "schema": {
                "$ref": "#/components/schemas/_classify_bind_Request"
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "{\n  \"guid\": \"59\",\n  \"classify\": \"qb43ps\"\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/_classify_bind_Response"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"retCode\": \"365\",\n  \"retMsg\": \"ldnj7o\",\n  \"tips\": \"h4muw0\",\n  \"data\": {\n    \"waring\": \"You may have used non-display generics.\"\n  }\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "update",
        "parameters": []
      }
    }
  },
  "info": {
    "title": "dm-metadata",
    "version": "1.0.0"
  }
}